feat: 적 구조물 공격 로직 및 구조물 HP 동기화 추가
- EnemyAI에 전방 BoxCast 기반 구조물 감지/공격 로직 추가(Structure 타겟팅)
- 구조물 HP 동기화를 BuildingHealthNet으로 분리하고 마스터 권위로 데미지/수리 적용 및 RPC 동기화
- 기존 StructureSelectable은 HP 접근을 BuildingHealthNet으로 위임하도록 정리(연관 코드 수정)
- 디버그를 위해 BoxCast 범위 Gizmos 시각화 추가

+++ Assets/Project/Scripts/Building/BuildingTypeSO.cs
@@ -13,8 +13,8 @@ public class BuildingTypeSO : ScriptableObject
 
     public bool allowRotate = true;
     public int cost = 0;
-    public int maxHp = 0;
+    public float maxHp = 0;
 
     public int repairCost => (int)(cost * 0.1);
-    public int repairAmount => (int)(maxHp * 0.5);
+    public float repairAmount => (int)(maxHp * 0.5);
 }
